Auto expand URL input
By using a contenteditable div instead of an input element, along with a
few other tweaks, the input width will now match whatever is pasted or
typed, and if it's longer than the screen width, it will wrap.

This way, someone doesn't have to jump to the beginning/end of a fixed
size input to check if their URL is correct.